requirements.txt
python-telegram-bot
feedparser
beautifulsoup4
requests
python-dateutil
